library type changes

+++ MDK-ARM/User/libraries/clib/includes/phaseLockedLoop.h
@@ -4,23 +4,23 @@
 
 typedef struct pll_parameters{
 
-    double d;
-    double df;
-    double dz;
-    double q;
-    double qf;
-    double qz;
-    double beta;
-    double Pout;
-    double Iout;
-    double PIout;
-    double theta;
-    double theta_comp;
+    float d;
+    float df;
+    float dz;
+    float q;
+    float qf;
+    float qz;
+    float beta;
+    float Pout;
+    float Iout;
+    float PIout;
+    float theta;
+    float theta_comp;
 		
 }pll_parameters;
 
 
-void PLL(double alpha,pll_parameters* pll);
+void PLL(float alpha,pll_parameters* pll);
 
 
 #endif
