+++ MDK-ARM/User/libraries/mlib/source/filtering.c
@@ -2,11 +2,11 @@
 #include "mlib.h"
 
 
-double filtering(double rtInput,  double *coeff, uint16_t coeffLength, double *zValues){
+float filtering(float rtInput,  float *coeff, uint16_t coeffLength, float *zValues){
 
 	uint16_t i;
-  double *z1_ptr,*z2_ptr,*coeff_ptr;
-  double output;
+  float *z1_ptr,*z2_ptr,*coeff_ptr;
+  float output;
 
 	z1_ptr=zValues; 		//background data
 	z2_ptr=z1_ptr; 		//data update
@@ -35,9 +35,9 @@ double filtering(double rtInput,  double *coeff, uint16_t coeffLength, double *z
 //1 	coeff is sequenced fof coefficents suxh --> b1,a1 in array
 //2.0 function invocation shown below  
 
-double fof_implementation(double rtInput,double* xz,double yz,  double *coeffs)
+float fof_implementation(float rtInput,float* xz,float yz,  float *coeffs)
 {
-	double y;
+	float y;
 	
 	y=(*coeffs)*(rtInput+(*xz))-(*(coeffs+1)*(yz));
 		
@@ -56,9 +56,9 @@ double fof_implementation(double rtInput,double* xz,double yz,  double *coeffs)
 //2.0 function invocation shown below  
 //2.1 "output=sos_implementation(input(real-time),output(lef-side),coeff_data(array_name),SOS2data(different for all parameters ));
 
-double sos_implementation(double x ,double yBack,   double *coeffs, sos_parameters *back){
+float sos_implementation(float x ,float yBack,   float *coeffs, sos_parameters *back){
 
-	double y;
+	float y;
 
 	
 	y=x*(*coeffs)	+	(back->xz1)	*(*(coeffs+1))		+	(back->xz2)*(*(coeffs+2))
